Move permission to own module
